SK + SG - Functional? Why yes!
We done with most stuff boys. Tragically, despite building this on state
of the art APIs, the APIs are meh. Sai and I handled all the errors, so
it works no matter what, (read as won't crash), but isn't always
optimal. We can go over what is left before the demo tomorrow!